Follow Tab + Organization + Others
Moved a lot of the table set up into a new file TableSetup. This is
loaded during the showing of page1 instead of on the start of the
program. That gives the app time to make the table so the table is no
longer generate in code dynamically. DataSets for events table can be
created before runtime because of this shift. Events table now contains
much more data from parse. Follow tab is created.